+++ src/components/Calendar/CalendarBox.css
@@ -1,6 +1,3 @@
-* {
-  /* font-family: ‘BookkMyungjo-Bd’; */
-}
 .react-calendar {
   top: 6rem;
   width: 35rem;
