+++ src/components/LoginModal/LoginModal.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import styles from "@components/LoginModal.module.css";
+import styles from "@components/LoginModal/LoginModal.module.css";
 import { IoMdCloseCircleOutline } from "react-icons/Io";
 import { useNavigate } from "react-router-dom";
